+++ HISTORY.rst
@@ -40,6 +40,7 @@ latest (unreleased)
 * BSSFL-521: Add 2nd sales person to keep track Phillip Ray sales
 * BSSFL-510: Proforma invoice
 * BSSFL-504: Improve email, add all followers in mailto
+* BSSFL-524: Stock picking smart button to repair
 
 **Bugfixes**